+++ src/microphysics/eos_phoebus/eos_phoebus.cpp
@@ -125,18 +125,7 @@ std::shared_ptr<StateDescriptor> Initialize(ParameterInput *pin) {
   }
   FillRealParams(pin, base_params, names);
 
-  // TODO(BRR) Remove this and use unit_conv and code_constants from phoebus package
-  params.Add("unit_conv", phoebus::UnitConversions(pin));
-  auto &unit_conv = params.Get<phoebus::UnitConversions>("unit_conv");
-  // phoebus::UnitConversions unit_conv(pin);
-
-  params.Add("code_constants", phoebus::CodeConstants(unit_conv));
-
-  // Store unit conversions mainly for dump file output
-  params.Add("length_unit", unit_conv.GetLengthCodeToCGS());
-  params.Add("time_unit", unit_conv.GetTimeCodeToCGS());
-  params.Add("mass_unit", unit_conv.GetMassCodeToCGS());
-  params.Add("temperature_unit", unit_conv.GetTemperatureCodeToCGS());
+  phoebus::UnitConversions unit_conv(pin);
 
   // modifiers
   /*
