+++ src/radiation/monte_carlo.cpp
@@ -67,9 +67,9 @@ TaskStatus MonteCarloSourceParticles(MeshBlock *pmb, MeshBlockData<Real> *rc,
   const Real &minx_k = pmb->coords.x3f(kb.s);
   auto geom = Geometry::GetCoordinateSystem(rc);
 
-  StateDescriptor *eos = pmb->packages.Get("eos").get();
-  auto &unit_conv = eos->Param<phoebus::UnitConversions>("unit_conv");
-  auto code_constants = CodeConstants(unit_conv);
+  auto &phoebus_pkg = pmb->packages.Get("phoebus");
+  auto &unit_conv = phoebus_pkg->Param<phoebus::UnitConversions>("unit_conv");
+  auto &code_constants = phoebus_pkg->Param<phoebus::CodeConstants>("code_constants");
 
   const Real d3x = dx_i * dx_j * dx_k;
 
@@ -383,9 +383,9 @@ TaskStatus MonteCarloTransport(MeshBlock *pmb, MeshBlockData<Real> *rc,
   auto &swarm_species = swarm->Get<int>("species").Get();
   auto swarm_d = swarm->GetDeviceContext();
 
-  StateDescriptor *eos = pmb->packages.Get("eos").get();
-  auto &unit_conv = eos->Param<phoebus::UnitConversions>("unit_conv");
-  auto code_constants = CodeConstants(unit_conv);
+  auto phoebus_pkg = pmb->packages.Get("phoebus");
+  auto &unit_conv = phoebus_pkg->Param<phoebus::UnitConversions>("unit_conv");
+  auto &code_constants = phoebus_pkg->Param<phoebus::CodeConstants>("code_constants");
 
   const Real h_code = code_constants.h();
   const Real mp_code = code_constants.mp();
